// src/modules/client/routes/enums/route.enum.ts
export enum ClientRouteEnum {
  ROOT = '/client',
  SIGN_IN = '/client/signin',
  SIGN_UP = '/client/signup',
  VERIFICATION = '/client/email-verify',
  SIGN_UP_LAST_STEP = '/client/signup-last-step',
  PASSWORD_RECOVERY = '/client/forgot-password',
  BOOKINGS = '/client/bookings',
  REBOOK_APPOINTMENT = '/client/bookings/rebook',
  CREATE_APPOINTMENT = '/client/bookings/new',
  REVIEWS = '/client/reviews',
  HELP = '/client/help',
}